web_search_agent/requirements.txt
agno
ddgs
groq
python-dotenv
streamlit